add staghorn_rrt script, revise blender assets directory
